server/requirements.txt
# ===== CORE FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
# uvloop==0.19.0       # Faster async event loop (not supported on Windows)
python-multipart==0.0.6

# ===== DATA VALIDATION =====
pydantic==2.5.0
pydantic[email]==2.5.0

# ===== ENVIRONMENT & CONFIGURATION =====
python-dotenv==1.0.0

# ===== DATABASE =====
pymongo==4.6.0

# ===== AUTHENTICATION & SECURITY =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# ===== AI/LLM PROVIDERS =====
anthropic==0.7.1          # Claude API
openai==1.3.7             # OpenAI GPT API
groq==0.4.1               # Groq Llama API

# ===== EXTERNAL API CLIENTS =====
requests==2.31.0          # HTTP requests
httpx==0.25.2             # Async HTTP

# ===== UTILITIES =====
numpy==1.26.2             # Numerical computing

# ===== TESTING & DEVELOPMENT =====
pytest==7.4.3
pytest-cov==4.1.0         # Coverage reporting
pytest-asyncio==0.21.1
pytest-mock==3.12.0       # Mock utilities
black==23.12.0             # Code formatter
flake8==6.1.0              # Linter
mypy==1.7.1                # Type checker
coverage==7.3.2            # Coverage analysis
